==> 02_Coordendas_CP.ipynb <==
## Obtener coordenadas de los codigos postales de España

Para obetener las coordenadas de los codigos postales  
Haremos uso de la API de googlemaps para obtener dichas coordenadas  
Para ello necesitaremos tener instaladas las siguientes paquetes y librerias:

#instalamos el paquete necesario para la API de googlemaps y las librerias necesarias.
!pip install googlemaps
import pandas as pd
import googlemaps

#Fuente externa. Fichero donde obtendremos los codigos postales de España, junto al tamaño del mercado del "Home Furnishing"
df=pd.read_excel('Datos_CP.xlsx')
df.head()

#Tratamos los datos co el formato 5 digitos del codigo postal y añadimos el pais, para evitar coincidencias de ubicaciones de 
#códigos postales en otros paises.
df['CP_Spain']= '0'+df['CP'].astype(str)
df['CP_Spain']= df['CP_Spain'].str[-5:]+' ,'+df['MUNICIPIO']+', '+df['PROVINCIA']+', Spain'
df= df[['CP_Spain','CP','Market Size']]
df.head()

#Ingresamos la key habilitada par conecatarnos a la API de googlemaps.
gmaps_key= googlemaps.Client(key='AIzaSyBbjRVDE0e0MyGlavAx0Zg-k-JKIUv2jng')

#Creamos las columnas necesarias para alojar los resultados recibido.
df['Latitude']= None
df['Longitude']= None

#Recorremos todo el listado de codigos postales y recibiendo sus coordenadas.
for i in range(0,len(df),1):
    geocode_result= gmaps_key.geocode(df.iat[i,0])
    try:
        lat= geocode_result[0]['geometry']['location']['lat']
        lon= geocode_result[0]['geometry']['location']['lng']
        df.iat[i, df.columns.get_loc('Latitude')] = lat
        df.iat[i, df.columns.get_loc('Longitude')] = lon
    except:
        lat = None
        lon = None

#Ordenamos el dataframe y eliminamos la columna "CP_Spain" que usamos para obtener las coordenadas
df= df[['CP','Market Size','Latitude','Longitude']]
df.head()

df['Market Size']=df['Market Size'].astype(int)
df['Latitude']=df['Latitude'].astype(float).round(6)
df['Longitude']=df['Longitude'].astype(float).round(6)

#Extraemos el dataframe a csv para su uso posterior.
df.to_csv('02_Output_Codigos_Postales.csv' ,sep=',',decimal=".",index=False)
