==> 03_CP20km.ipynb <==
# Calcular todos los CP a menos de una distacia máxima para todos los CP de España

import pandas as pd
import math

zipc = pd.read_csv('02_Output_Codigos_Postales.csv', dtype={'CP':'str', 'Latitude':'float','Longitude':'float'})
maxdist = 20

df = pd.DataFrame(columns=['CP', 'CP2', 'Dist','Market Size'])
line=1

for cp1,row in zipc.iterrows():
    cpa = row[0]
    lata = row[2]
    lona = row[3]
    if cp1 % 500 == 0:
        print(cp1)
    for cp2,row in zipc.iterrows():
        cpb = row[0]
        latb = row[2]
        lonb = row[3]
        msb = row[1]
        try:
            dist = math.acos(math.cos(math.radians(90-lata)) * math.cos(math.radians(90-latb)) + math.sin(math.radians(90-lata)) * math.sin(math.radians(90-latb)) * math.cos(math.radians(lona-lonb))) * 6371
        except:
            dist = 0
        if dist <= maxdist:
            df.loc[line] = (cpa,cpb,dist,msb)
            line +=1 

            
# Exportamos el resultado a un CSV
df.to_csv('03_Output_CP20km.csv', sep=',',decimal=".", index=False)
