==> 04_Diseña_sales.ipynb <==
# Ventas de las tiendas vs distancia

### Código todo en uno ejecutable

import pandas as pd
import math

# Importamos las ventas por CP de nuestras tiendas (incluye nombre de la tienda, CP de la tienda, CP de la venta, cantidad vendida):
sales = pd.read_csv('Sales_per_point.csv',sep=",",decimal=".")

# Importamos las coordenadas de todos los CP de España:
coord = pd.read_csv('02_Output_Codigos_Postales.csv',sep=",",decimal=".", dtype={'CP':'int', 'Market Size':'float','Latitude':'float','Longitude':'float'})

# Incluimos para cada CP de la tienda y CP de venta sus coordenadas:
sales2 = sales.merge(coord, left_on='CP CMP', right_on='CP', how='left')
sales2.rename(columns={'Latitude':'lata', 'Longitude':'lona'},inplace=True)
sales2.drop(['CP','Market Size'], inplace=True, axis=1)

df = sales2.merge(coord, left_on='CP Sales', right_on='CP', how='left')
df.rename(columns={'Latitude':'latb', 'Longitude':'lonb'},inplace=True)
df.drop('CP', inplace=True, axis=1)

# Calculamos la distancia entre el CP de la tienda y de la venta basándonos en las coordenadas:
df2 = pd.DataFrame(columns=['CMP', 'CP1', 'CP2', 'Dist', 'Roundist', 'Sales', 'Market Size'])
line=1

for cp,row in df.iterrows():
    cmp = row[1]
    cp1 = row[2]
    cp2 = row[3]
    sales = row[4]
    ms = row[7]
    lata = row[5]
    lona = row[6]
    latb = row[8]
    lonb = row[9]
    try:
        dist = math.acos(math.cos(math.radians(90-lata)) * math.cos(math.radians(90-latb)) + math.sin(math.radians(90-lata)) * math.sin(math.radians(90-latb)) * math.cos(math.radians(lona-lonb))) * 6371
    except:
        dist = 0
    rdist = round(dist)
    df2.loc[line] = (cmp, cp1, cp2, dist, rdist, sales, ms)
    line +=1 
    
# Agrupamos las ventas por distancia
Dist_sales = df2.groupby(['CMP','Roundist'],as_index=False).agg({'Sales':'sum', 'Market Size':'sum'}) 

#Añadimos las ventas acumuladas y el % que supone del total. Eliminamos aquellos CP demasiado alejados
Dist_sales['Pct'] = (Dist_sales['Sales']/Dist_sales['Sales'].sum())
Dist_sales['Acc_sales'] = Dist_sales['Sales'].cumsum()
Dist_sales['Acc_Pct'] = (Dist_sales['Acc_sales']/Dist_sales['Sales'].sum())
Dist_sales['Acc_MSize'] = Dist_sales['Market Size'].cumsum()
Dist_sales['MShare'] = Dist_sales['Sales']/Dist_sales['Market Size']
Dist_sales['Acc_MShare'] = Dist_sales['Acc_sales']/Dist_sales['Acc_MSize']

CMP_Dist_Sales = Dist_sales[Dist_sales.Roundist < 21]

# Exportamos el resultado a un CSV
CMP_Dist_Sales.to_csv('04_Output_CMP_Dist_sales.csv', sep=',',decimal=".", index=False)

# Exportamos las ventas y Market Share actuales a un CSV
CP_MShare = df2.groupby(['CP2'],as_index=False).agg({'Sales':'sum', 'Market Size':'mean'})
CP_MShare['Actual_MarketShare'] = CP_MShare['Sales']/CP_MShare['Market Size']
CP_MShare.drop(['Sales','Market Size'], inplace=True, axis=1)
CP_MShare.to_csv('04_Output_CP_Actual_MarketShare.csv', sep=',',decimal=".", index=False)

### Ejercicio por partes

Juntamos las ventas por CP por cada Punto de venta y las coordenadas, para saber la venta por distancia y el tamaño del mercado

import pandas as pd
import math

# Importamos las ventas por CP de nuestras tiendas (incluye nombre de la tienda, CP de la tienda, CP de la venta, cantidad vendida):
sales = pd.read_csv('Sales_per_point.csv',sep=",",decimal=".")

# Importamos las coordenadas de todos los CP de España:
coord = pd.read_csv('02_Output_Codigos_Postales.csv',sep=",",decimal=".", dtype={'CP':'int', 'Market Size':'float','Latitude':'float','Longitude':'float'})

# Incluimos para cada CP de la tienda y CP de venta sus coordenadas:
sales2 = sales.merge(coord, left_on='CP CMP', right_on='CP', how='left')
sales2.rename(columns={'Latitude':'lata', 'Longitude':'lona'},inplace=True)
sales2.drop(['CP','Market Size'], inplace=True, axis=1)

df = sales2.merge(coord, left_on='CP Sales', right_on='CP', how='left')
df.rename(columns={'Latitude':'latb', 'Longitude':'lonb'},inplace=True)
df.drop('CP', inplace=True, axis=1)

# Calculamos la distancia entre el CP de la tienda y de la venta basándonos en las coordenadas:
df2 = pd.DataFrame(columns=['CMP', 'CP1', 'CP2', 'Dist', 'Roundist', 'Sales', 'Market Size'])
line=1

for cp,row in df.iterrows():
    cmp = row[1]
    cp1 = row[2]
    cp2 = row[3]
    sales = row[4]
    ms = row[7]
    lata = row[5]
    lona = row[6]
    latb = row[8]
    lonb = row[9]
    try:
        dist = math.acos(math.cos(math.radians(90-lata)) * math.cos(math.radians(90-latb)) + math.sin(math.radians(90-lata)) * math.sin(math.radians(90-latb)) * math.cos(math.radians(lona-lonb))) * 6371
    except:
        dist = 0
    rdist = round(dist)
    df2.loc[line] = (cmp, cp1, cp2, dist, rdist, sales, ms)
    line +=1 
df2.head()

Comprobaciones. Distancia a la que vendemos >90%, y el MSize caiga por debajo del 5%

# Agrupamos las ventas por distancia
Dist_sales = df2.groupby(['Roundist'],as_index=False).agg({'Sales':'sum', 'Market Size':'sum'}) 

#Añadimos las ventas acumuladas y el % que supone del total
Dist_sales['Pct'] = (Dist_sales['Sales']/Dist_sales['Sales'].sum())
Dist_sales['Acc_sales'] = Dist_sales['Sales'].cumsum()
Dist_sales['Acc_Pct'] = (Dist_sales['Acc_sales']/Dist_sales['Sales'].sum())
Dist_sales['Acc_MSize'] = Dist_sales['Market Size'].cumsum()
Dist_sales['MShare'] = Dist_sales['Sales']/Dist_sales['Market Size']
Dist_sales['Acc_MShare'] = Dist_sales['Acc_sales']/Dist_sales['Acc_MSize']

Dist_sales.head(40)

Exportamos para el siguiente paso

# Agrupamos las ventas por distancia
Dist_sales = df2.groupby(['CMP','Roundist'],as_index=False).agg({'Sales':'sum', 'Market Size':'sum'}) 

#Añadimos las ventas acumuladas y el % que supone del total
Dist_sales['Pct'] = (Dist_sales['Sales']/Dist_sales['Sales'].sum())
Dist_sales['Acc_sales'] = Dist_sales['Sales'].cumsum()
Dist_sales['Acc_Pct'] = (Dist_sales['Acc_sales']/Dist_sales['Sales'].sum())
Dist_sales['Acc_MSize'] = Dist_sales['Market Size'].cumsum()
Dist_sales['MShare'] = Dist_sales['Sales']/Dist_sales['Market Size']
Dist_sales['Acc_MShare'] = Dist_sales['Acc_sales']/Dist_sales['Acc_MSize']

CMP_Dist_Sales = Dist_sales[Dist_sales.Roundist < 21]
CMP_Dist_Sales.head(30)

# Exportamos el resultado a un CSV
CMP_Dist_Sales.to_csv('04_Output_CMP_Dist_sales.csv', sep=',',decimal=".", index=False)

Exportamos para el último paso

CP_MShare = df2.groupby(['CP2'],as_index=False).agg({'Sales':'sum', 'Market Size':'mean'})
CP_MShare['Actual_MarketShare'] = CP_MShare['Sales']/CP_MShare['Market Size']
CP_MShare.drop(['Sales','Market Size'], inplace=True, axis=1)
CP_MShare.to_csv('04_Output_CP_Actual_MarketShare.csv', sep=',',decimal=".", index=False)
