==> 06_Regression_CP_España.ipynb <==
# Analisis del comportamiento de la cuota de mercado vs distancia

Queremos predecir las ventas de una tienda basándonos en la cuota de mercado que alcanza de un CP dependiendo de la distancia a ese CP.

### Código todo en uno ejecutable

# Importar el modelo
import pandas as pd
import pickle
DT_model = pickle.load(open("05_Output_Regression_DT.pickle","rb"))

# Importar distancias entre CP y market size
zipcodes = pd.read_csv("03_Output_CP20km.csv")
zipcodes['Roundist'] = round (zipcodes['Dist'])

# Aplicamos el modelo para predecir el % Market share dependiendo de la distancia y lo multiplicamos por el Market Size para saber la predicción de ventas
zipcodes['MShare_Prediction'] = DT_model.predict(zipcodes[['Roundist']])
zipcodes['Sales_Prediction'] = zipcodes['Market Size'] * zipcodes['MShare_Prediction']

### Restamos de la predicción de Market Share el Market Share actual al que llegamos con nuestras tiendas, para quedarnos sólo con zonas potenciales

# Importamos el Market Share que tenemos con las tiendas actuales
actsales = pd.read_csv("04_Output_CP_Actual_MarketShare.csv")

# Calculamos los potenciales
potentials = zipcodes.merge(actsales, on='CP2', how='left').fillna(0)
potentials['Potential_Share'] = potentials['MShare_Prediction'] - potentials['Actual_MarketShare']
potentials.loc[potentials['Potential_Share'] < 0, 'Potential_Share'] = 0 
potentials['Potential_Sales'] = potentials['Market Size'] * potentials['Potential_Share']

potentials.to_csv('06_Output_Sales_Prediction.csv', sep=',',decimal=".", index=False)

--------

### Ejercicio por partes

# Importar el modelo
import pickle
DT_model = pickle.load(open("05_Output_Regression_DT.pickle","rb"))

import pandas as pd

# Importar distancias entre CP y market size
zipcodes = pd.read_csv("03_Output_CP20km.csv")
zipcodes['Roundist'] = round (zipcodes['Dist'])

# Aplicamos el modelo para predecir el % Market share dependiendo de la distancia y lo multiplicamos por el Market Size para saber la predicción de ventas
zipcodes['MShare_Prediction'] = DT_model.predict(zipcodes[['Roundist']])
zipcodes['Sales_Prediction'] = zipcodes['Market Size'] * zipcodes['MShare_Prediction']

zipcodes.head()

### Restamos de la predicción de Market Share el Market Share actual al que llegamos con nuestras tiendas, para quedarnos sólo con zonas potenciales
# Importamos el Market Share que tenemos con las tiendas actuales
actsales = pd.read_csv("04_Output_CP_Actual_MarketShare.csv")

# Calculamos los potenciales
potentials = zipcodes.merge(actsales, on='CP2', how='left').fillna(0)
potentials['Potential_Share'] = potentials['MShare_Prediction'] - potentials['Actual_MarketShare']
potentials.loc[potentials['Potential_Share'] < 0, 'Potential_Share'] = 0 
potentials['Potential_Sales'] = potentials['Market Size'] * potentials['Potential_Share']
potentials.head()

Dist_sales = potentials.groupby(['CP'],as_index=False).agg({'Potential_Sales':'sum'})
Dist_sales.sort_values(by=['Potential_Sales'], ascending=False, inplace=True)
Dist_sales.head()
